requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0